requirements.txt
captum==0.7.0
grad_cam==1.4.6
matplotlib==3.7.0
multiprocess==0.70.16
numpy==2.2.4
opencv_python==4.5.5.64
pandas==2.2.3
Pillow==8.2.0
Pillow==11.2.1
pytorch_lightning==2.4.0
scipy==1.15.2
torch==2.4.1+cu124
torch_dct==0.1.6
torchvision==0.19.1+cu124
tqdm==4.64.1
tqdm==4.67.1
trulens==0.13.4
